requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-extensions==4.1
djangorestframework==3.16.1
idna==3.10
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
